+++ index.html
@@ -5,9 +5,7 @@
 	<meta name="viewport" content="width=device-width, initial-scale=1">
 	<title>News</title>
 	<link rel="stylesheet" href="styles.css">
-	<script src="//cdn.jsdelivr.net/bluebird/3.5.0/bluebird.min.js"></script>
 	<script src="https://cdnjs.cloudflare.com/ajax/libs/fetch/2.0.3/fetch.js"></script>
-	<script src="https://cdnjs.cloudflare.com/ajax/libs/babel-core/5.6.15/browser-polyfill.min.js"></script>
 </head>
 <body>
 	<aside id="side-block" class="collapsed">
@@ -38,6 +36,6 @@
 		<div class="created">Created using <a href="https://newsapi.org/" target="_blank">NewsAPI</a></div>
 	</main>
 
-	<script src="app/index.js"></script>
+	<script src="dist/main.js"></script>
 </body>
 </html>
