+++ package.json
@@ -2,8 +2,8 @@
   "name": "yurybolbas-fc.github.io",
   "version": "1.0.0",
   "scripts": {
-    "build": "babel src -d app",
-    "watch": "babel src --watch -d app"
+    "dev": "webpack --mode development",
+    "prod": "webpack --mode production"
   },
   "repository": {
     "type": "git",
@@ -15,7 +15,15 @@
     "@babel/plugin-transform-typeof-symbol": "^7.0.0",
     "@babel/polyfill": "^7.0.0",
     "@babel/preset-env": "^7.1.6",
-    "babel-plugin-transform-es2015-typeof-symbol": "^6.23.0"
+    "babel-loader": "^8.0.4",
+    "babel-plugin-transform-es2015-typeof-symbol": "^6.23.0",
+    "webpack": "^4.26.0",
+    "webpack-cli": "^3.1.2",
+    "webpack-dev-server": "^3.1.10"
   },
-  "dependencies": {}
+  "dependencies": {
+    "babel-polyfill": "^6.26.0",
+    "browser-polyfill": "^3.20.2",
+    "fetch-polyfill": "^0.8.2"
+  }
 }
